+++ src/main/java/it/polimi/ingsw/model/player/iSpaceShip.java
@@ -26,7 +26,7 @@ public interface iSpaceShip {
 
 	GameModeType getType();
 
-	float getCannonPower();
+	double getCannonPower();
 
 	int[] getCrew();
